Separated the processor and callback into two python files.
